research/ProtienClusterPrediction/protein-transformer-clustering/pyproject.toml
[tool.poetry]
name = "protein-transformer-clustering"
version = "0.1.0"
description = "A project to cluster protein sequences based on resistance patterns using a transformer model."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.9.0"
numpy = "^1.21.0"
scikit-learn = "^0.24.0"
pyyaml = "^5.4.1"
pandas = "^1.3.0"
matplotlib = "^3.4.0"
seaborn = "^0.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"